Core big integers arrive typed in both apps

+++ android/data/coordinators/src/main/kotlin/com/gemwallet/android/data/coordinators/banner/GetActiveBannersImpl.kt
@@ -62,6 +62,6 @@ class GetActiveBannersImpl(
 
     private fun hasStakeBalance(assetInfo: AssetInfo?): Boolean {
         val balance = assetInfo?.balance ?: return false
-        return balance.toGem().stakedValue(assetInfo.asset.chain.string).toBigInteger() > BigInteger.ZERO
+        return balance.toGem().stakedValue(assetInfo.asset.chain.string) > BigInteger.ZERO
     }
 }
